fix: new network interface, todo refactor BinaryStream to ByteBuf

+++ src/main/java/cn/nukkit/network/protocol/AgentAnimationPacket.java
@@ -14,7 +14,7 @@ public class AgentAnimationPacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.animation = (byte) getByte();
+        this.animation = getByte();
         this.runtimeEntityId = getEntityRuntimeId();
     }
